atd-toolbox/isolate_vz_data_differences/requirements.txt
psycopg2-binary
visidata
tqdm
